search-bar on every page
