+++ src/network.css
@@ -191,7 +191,7 @@ main {
   border-radius: 15px;
   background-color: white;
   margin-top: 2%;
-  width: 88.75%;
+  width: 100%;
   word-wrap: break-word;
   height: 100%;
   padding-top: 2%;
