+++ src/pages/network.js
@@ -7,17 +7,7 @@ import Profile from './profile';
 import Error from './error';
 import Edit from './edit';
 import Post from './post';
-
-function SearchBar(){
-  var val = document.getElementById("search-input").value;
-  var opts = document.getElementById('search-bar').childNodes;
-  for (var i = 0; i < opts[0].childNodes.length; i++) {
-    if (opts[0].childNodes[i].value === val) {
-      window.location.href=("/profile/" + opts[0].childNodes[i].value)
-      break;
-    }
-  }
-}
+import Header  from '../components/header';
 
 function SocialNetwork(props) {
 
@@ -30,15 +20,7 @@ function SocialNetwork(props) {
       <Routes>
         <Route  path="/" element={
         <body>
-        <header className="App-header" id="App-header">
-            <div className='inner-header'>
-              <a href=''><h3>Social Network</h3></a>
-              <div className='header-search'><h3>
-                <input list='search-bar' id='search-input' type="text" placeholder="Suche..." className='search-bar' onInput={SearchBar}/>
-                <datalist id="search-bar"></datalist>
-                </h3></div>
-            </div>
-        </header>
+        <Header />
         <main>
           <div className='feed' id='feed' >
           </div>
